// apps/www/src/types.ts
export type Site = {
  title: string
  description: string
  href: string
  author: string
  locale: string
  featuredProjectCount: number
  featuredPublicationCount: number
  pageSize: number
}

export type SocialLink = {
  href: string
  label: string
  icon?: string
}

export type IconMap = {
  [key: string]: string
}